refactor: movimentação das configurações do axios
- as confiurações do axios foram movidas para a pasta lib
- a pasta lib fica responsável por armazenar configurações das
  bibliotecas usadas ao longo do projeto"
