+++ frontend/src/screens/auth/SignupScreen.tsx
@@ -7,7 +7,7 @@ import { ScreenWrapper } from "../../components/ScreenWrapper";
 import { Input } from "../../components/Input";
 import { PasswordInput } from "../../components/PasswordInput";
 import { SubmitButton } from "../../components/buttons/SubmitButton";
-import { request } from "../../services/Request";
+import { request } from "../../libs/axios/Request";
 import { ScreenTitle } from "../../components/ScreenTitle";
 
 const SignupScreen = ({ navigation }: { navigation: NavigationProp<any> }) => {
